Move sprites into their own source file.
